check_cpu_config/requirements.txt
psutil
py-cpuinfo
pandas